requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
pycparser==2.22
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1